added UI changes according to gif flow
