+++ likeminds-chat-reactnative-integration/ChatSX/components/ImageVideoConversation/index.tsx
@@ -58,7 +58,7 @@ const ImageVideoConversation = () => {
   const lmChatInterface = CallBack.lmChatInterface;
 
   return (
-    <>
+    <View style={styles.messageParent}>
       <View
         style={[
           styles.displayRow,
@@ -188,7 +188,7 @@ const ImageVideoConversation = () => {
           )}
         </View>
       ) : null}
-    </>
+    </View>
   );
 };
