+++ likeminds-chat-reactnative-integration/ChatSX/components/ReplyConversations/index.tsx
@@ -292,6 +292,7 @@ const ReplyConversations = () => {
         {
           justifyContent: isTypeSent ? "flex-end" : "flex-start",
         },
+        styles.messageParent,
       ]}
     >
       <View
